MHD-014 Add unix openGL and simulation support

+++ include/openGL/IndexBuffer.h
@@ -3,7 +3,7 @@
 class IndexBuffer {
 private:
     unsigned int m_RendererID;
-    size_t m_Count;
+    unsigned int m_Count;
 
 public:
     IndexBuffer(const unsigned int* data, unsigned int count);
@@ -12,5 +12,5 @@ public:
     void Bind() const;
     void Unbing() const;
 
-    inline size_t GetCount() const { return m_Count; }
+    inline unsigned int GetCount() const { return m_Count; }
 };
